Model DAO implemented
I added a model DAO in order to used the fetched data from the
NetworkDAO and populate our model classes. For the moment only the code
for PersonalInformation has been written. The rest will come when the
code proved to be working.

In order to have the class populated, I added a JSON dependency to the
pom.xml

There were a few typos in PersonalInformation, they have been corrected.