Added option to update the existing data
I created the method findExistingPerson() in order to be able to know if
an id already exists. From this knowledge, I was able to update the
showPersonInfoForm() and submitPersonInfoForm() in order to recognise if
the set of data already exists and update it. If it does not, the
program will just run as usual and add the data to the existing one.