I am back from holidays! I missed coding sooo much!
Ok! I had some real problem witht he validation annotations! but I
finally found the solution. I added the annotation @ModelAttribute to
the controller method in order to implement the Data binding from MVC
Spring.

Also, my big mistake was in the HTML! HTML is so easy though, but I
guess it is also easy to do a mistake. The span in which I added the
Thymeleaf annotation th:errors was outside the input div! I fixed that
for all the fields annnnnd Tada! Problem solved